+++ husky_components/dropzone/main.js
@@ -214,6 +214,15 @@ define([], function() {
             return createEventName.call(this, 'disable');
         },
 
+        /**
+         * listens on and adds an image to the dropzone by a given url
+         * @event husky.dropzone.<instance-name>.add-file
+         * @param {String} url The url to the file to add
+         */
+        ADD_IMAGE = function() {
+            return createEventName.call(this, 'add-image');
+        },
+
         /** returns normalized event names */
         createEventName = function(postFix, global) {
             return [
@@ -252,13 +261,15 @@ define([], function() {
 
             this.sandbox.emit(INITIALIZED.call(this));
         },
-        
+
         /**
          * Handler which gets called when the component gets destroyed.
          */
         destroy: function() {
             this.sandbox.stop(this.$loader);
             $('body').off('.dropzone' + this.options.instanceName);
+            this.dropzone.destroy();
+            this.$dropzone.remove();
         },
 
         /**
@@ -329,6 +340,24 @@ define([], function() {
                     this.openOverlay();
                 }.bind(this));
             }
+
+            this.sandbox.on(ADD_IMAGE.call(this), this.addImage.bind(this));
+        },
+
+        /**
+         * Adds an image to the the dropzone by a given url.
+         *
+         * @param {String} url The url to first load the image from and than upload it via the dropzone
+         * @param {string} mimeType The mime type of the loaded image
+         * @param {string} fileName The name under which the file should be saved
+         */
+        addImage: function(url, mimeType, fileName) {
+            this.sandbox.util.loadImageAsBlob(url, mimeType).then(function(imageBlob) {
+                if (!!fileName) {
+                    imageBlob.name = fileName;
+                }
+                this.dropzone.addFile(imageBlob, mimeType);
+            }.bind(this));
         },
 
         /**