+++ husky_services/logger.js
@@ -1,0 +1,20 @@
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+define(function() {
+
+    'use strict';
+
+    function Logger() {
+    }
+
+    Logger.prototype = window.Husky.logger;
+
+    return new Logger();
+});